change the local storage to use the pg DB also make improvments to the application, change the file structre to a cleaner code
