+++ .env.example
@@ -1,3 +1,4 @@
 PORT=
 SPOONACULAR_API_KEY=
 BASE_URL=
+DATABASE_URL=
