+++ public/randomRecipes.html
@@ -5,7 +5,8 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Flavor Table — Random Recipe</title>
     <link rel="stylesheet" href="./styles.css" />
-    <script src="./app.js" defer></script>
+    <script src="https://cdn.jsdelivr.net/npm/axios/dist/axios.min.js"></script>
+    <script type="module" src="./app.js" defer></script>
   </head>
   <body>
     <header id="main-header">
@@ -13,8 +14,8 @@
         <h1 id="brand">Flavor Table</h1>
         <ul id="nav-links">
           <li><a href="/">Home</a></li>
-          <li><a href="./randomRecipes.html">Random Recipe</a></li>
           <li><a href="./favorites.html">Favorites</a></li>
+          <li><a href="./randomRecipes.html">Random Page</a></li>
         </ul>
       </nav>
     </header>
@@ -39,8 +40,7 @@
         <p>Made with ❤️ by Saif Al Kurdi, 2025.</p>
         <nav id="footer-nav">
           <a href="/">Home</a> | <a href="./randomRecipes.html">Random</a> |
-                    <a href="./favorites.html">Favorites
-
+          <a href="./favorites.html">Favorites</a>
         </nav>
         <p id="copyright">&copy; 2025 Flavor Table. All rights reserved.</p>
       </div>
